+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Auth shared `URLSessionClient` defaults to background operation queue.
 - Remove `XMLNodeIndex`.
 - Update `XMLNode` subscript method to return `XMLNode?`.
+- Add `.firstLevel` traversal for `XMLNode` to retrieve direct children only.
 
 #### Enhancements
 - Shared `URLSessionClient` with default background operation queue.